address_review_comments

+++ spark/src/test/scala/org/apache/comet/CometCastSuite.scala
@@ -648,7 +648,7 @@ class CometCastSuite extends CometTestBase with AdaptiveSparkPlanHelper {
   }
 
   test("cast StringType to FloatType") {
-    castTest(gen.generateStrings(dataSize, numericPattern, 8).toDF("a"), DataTypes.DoubleType)
+    castTest(gen.generateStrings(dataSize, numericPattern, 8).toDF("a"), DataTypes.FloatType)
   }
 
   val specialValues: Seq[String] = Seq(
